Make `gateway` a module as there was too much in `gateway.rs`.
